requirements.txt
torch
torchvision
torchaudio
transformers
peft
trl
bitsandbytes
accelerate
datasets
wandb
spacy
openai
python-dotenv
